Fixed Collision with invisible blocks, fixed coordinate data type

+++ libexodus/src/movement.rs
@@ -1,5 +1,5 @@
 #[derive(Copy, Clone, PartialEq)]
 pub struct Movement {
     pub velocity: (f32, f32),
-    pub target: (f32, f32),
+    pub target: (i32, i32),
 }
